requirements.txt
adal==1.2.5
alembic==1.4.1
altair==4.1.0
ansiwrap==0.8.4
applicationinsights==0.11.9
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
azure-common==1.1.26
azure-core==1.9.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.3.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.6.0
azure-storage-queue==12.1.4
azureml-automl-core==1.18.0.post1
azureml-automl-runtime==1.18.0.post2
azureml-contrib-notebook==1.18.0
azureml-core==1.18.0.post2
azureml-dataprep==2.4.4
azureml-dataprep-native==24.0.0
azureml-dataprep-rslex==1.2.3
azureml-dataset-runtime==1.18.0
azureml-defaults==1.18.0
azureml-interpret==1.18.0
azureml-model-management-sdk==1.0.1b6.post1
azureml-pipeline==1.18.0
azureml-pipeline-core==1.18.0
azureml-pipeline-steps==1.18.0
azureml-sdk==1.18.0
azureml-telemetry==1.18.0
azureml-train==1.18.0
azureml-train-automl==1.18.0
azureml-train-automl-client==1.18.0
azureml-train-automl-runtime==1.18.0.post1
azureml-train-core==1.18.0.post1
azureml-train-restclients-hyperdrive==1.18.0
azureml-widgets==1.18.0.post1
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.1.0
bayesian-optimization==1.2.0
bleach==3.2.1
blinker==1.4
blis==0.7.4
boto==2.49.0
boto3==1.15.18
botocore==1.18.18
cachetools==4.2.0
catalogue==1.0.0
catboost==0.24.3
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorlover==0.3.0
combo==0.1.1
configparser==3.7.4
confuse==1.4.0
contextlib2==0.6.0.post1
cryptography==3.2.1
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
databricks-cli==0.14.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distro==1.5.0
docker==4.3.1
docopt==0.6.2
dotnetcore2==2.1.19
entrypoints==0.3
et-xmlfile==1.0.1
fire==0.3.1
Flask==1.0.3
funcy==1.15
fusepy==3.0.1
future==0.18.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.11
graphviz==0.15
gunicorn==19.9.0
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
imbalanced-learn==0.7.0
importlib-metadata==2.0.0
iniconfig==1.1.1
interpret-community==0.15.4
interpret-core==0.2.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
jeepney==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json-logging-py==0.2
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
keras2onnx==1.6.0
kiwisolver==1.3.1
kmodes==0.10.2
liac-arff==2.5.0
lightgbm==3.1.1
llvmlite==0.35.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
missingno==0.4.2
mistune==0.8.4
mlflow==1.12.1
mlxtend==0.18.0
msal==1.6.0
msal-extensions==0.2.2
msrest==0.6.19
msrestazure==0.6.4
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
ndg-httpsclient==0.5.1
nest-asyncio==1.4.3
networkx==2.5
nimbusml==1.8.0
nltk==3.5
notebook==6.1.5
numba==0.52.0
numexpr==2.7.1
numpy==1.19.4
oauthlib==3.1.0
onnx==1.7.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
onnxruntime==1.3.0
openpyxl==3.0.5
packaging==20.4
pandas==1.2.0
pandas-profiling==2.9.0
pandocfilters==1.4.3
papermill==1.2.1
parso==0.7.1
pathspec==0.8.1
patsy==0.5.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==8.0.1
pipreqs==0.4.10
plac==1.1.3
plotly==4.14.1
pluggy==0.13.1
pmdarima==1.1.1
portalocker==1.7.1
preshed==3.0.5
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.7.3
ptyprocess==0.6.0
py==1.9.0
py-cpuinfo==5.0.0
pyarrow==1.0.1
pyasn1==0.4.8
pycaret==2.2.2
pycodestyle==2.6.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.2
PyJWT==1.7.1
pyLDAvis==2.1.2
pyod==0.8.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
querystring-parser==1.2.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.0
SecretStorage==3.2.0
Send2Trash==1.5.0
shap==0.34.0
six==1.15.0
skl2onnx==1.4.9
sklearn-pandas==1.7.0
smart-open==1.9.0
smmap==3.0.4
spacy==2.3.4
SQLAlchemy==1.3.20
sqlparse==0.4.1
srsly==1.0.5
statsmodels==0.10.2
streamlit==0.74.1
suod==0.0.4
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
tenacity==6.2.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
textwrap3==0.9.2
thinc==7.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
umap-learn==0.4.6
urllib3==1.26.2
validators==0.18.2
visions==0.5.0
wasabi==0.8.0
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
xgboost==1.2.1
xlrd==1.2.0
yarg==0.1.9
yellowbrick==1.2
zipp==3.4.0
